Remove Simulator class
